+++ crates/tau-runtime/src/lib.rs
@@ -1,5 +1,9 @@
 pub mod channel_store;
+pub mod rpc_capabilities_runtime;
+pub mod rpc_protocol_runtime;
 pub mod transport_health;
 
 pub use channel_store::*;
+pub use rpc_capabilities_runtime::*;
+pub use rpc_protocol_runtime::*;
 pub use transport_health::*;
