+++ crates/tau-runtime/testdata/rpc-schema-compat/dispatch-unsupported-continues.json
@@ -1,0 +1,33 @@
+{
+  "schema_version": 1,
+  "name": "dispatch-unsupported-continues",
+  "mode": "dispatch_ndjson",
+  "input_lines": [
+    "{\"schema_version\":99,\"request_id\":\"req-bad-schema\",\"kind\":\"run.cancel\",\"payload\":{\"run_id\":\"run-legacy\"}}",
+    "{\"schema_version\":1,\"request_id\":\"req-good-start\",\"kind\":\"run.start\",\"payload\":{\"prompt\":\"hello\"}}"
+  ],
+  "expected_processed_lines": 2,
+  "expected_error_count": 1,
+  "expected_responses": [
+    {
+      "schema_version": 1,
+      "request_id": "req-bad-schema",
+      "kind": "error",
+      "payload": {
+        "code": "unsupported_schema",
+        "message": "unsupported rpc frame schema: supported request schema versions are [0, 1], found 99"
+      }
+    },
+    {
+      "schema_version": 1,
+      "request_id": "req-good-start",
+      "kind": "run.accepted",
+      "payload": {
+        "status": "accepted",
+        "mode": "preflight",
+        "prompt_chars": 5,
+        "run_id": "run-req-good-start"
+      }
+    }
+  ]
+}
