+++ crates/tau-runtime/testdata/rpc-schema-compat/serve-unsupported-continues.json
@@ -1,0 +1,56 @@
+{
+  "schema_version": 1,
+  "name": "serve-unsupported-continues",
+  "mode": "serve_ndjson",
+  "input_lines": [
+    "{\"schema_version\":99,\"request_id\":\"req-bad-schema\",\"kind\":\"run.cancel\",\"payload\":{\"run_id\":\"run-missing\"}}",
+    "{\"schema_version\":1,\"request_id\":\"req-start\",\"kind\":\"run.start\",\"payload\":{\"prompt\":\"x\"}}"
+  ],
+  "expected_processed_lines": 2,
+  "expected_error_count": 1,
+  "expected_responses": [
+    {
+      "schema_version": 1,
+      "request_id": "req-bad-schema",
+      "kind": "error",
+      "payload": {
+        "code": "unsupported_schema",
+        "message": "unsupported rpc frame schema: supported request schema versions are [0, 1], found 99"
+      }
+    },
+    {
+      "schema_version": 1,
+      "request_id": "req-start",
+      "kind": "run.accepted",
+      "payload": {
+        "status": "accepted",
+        "mode": "preflight",
+        "prompt_chars": 1,
+        "run_id": "run-req-start"
+      }
+    },
+    {
+      "schema_version": 1,
+      "request_id": "req-start",
+      "kind": "run.stream.tool_events",
+      "payload": {
+        "run_id": "run-req-start",
+        "event": "run.started",
+        "mode": "preflight",
+        "sequence": 0
+      }
+    },
+    {
+      "schema_version": 1,
+      "request_id": "req-start",
+      "kind": "run.stream.assistant_text",
+      "payload": {
+        "run_id": "run-req-start",
+        "delta": "preflight run accepted (1 prompt chars)",
+        "final": false,
+        "mode": "preflight",
+        "sequence": 1
+      }
+    }
+  ]
+}
